uv: move warning from run to sync
